Move lin_op_helpers back to lin_op
